Abstract away multiple vs single SPA generation
